legacy/Nado_grid_bot/requirements.txt
websockets>=12.0
pyyaml>=6.0
python-dotenv>=1.0
